Added Page  object class + new test class
Added cool features
