+++ PageObjects/SamplePageObjects.py
@@ -1,0 +1,3 @@
+
+def login_text_header():
+    return"//h5[@class='modal-title title'][text()='Login']"
